Improve codebase architecture
